+++ mealpy/human_based/MGOA.py
@@ -15,8 +15,8 @@ class OriginalMGOA(Optimizer):
     Links:
         1. https://doi.org/10.3390/sym17122118
 
-    Notes:
-        + Paper parameters: A = pi, n = 2, alpha = 1
+    Hyper-parameters should fine-tune in approximate range to get faster convergence toward the global optimum:
+        + attract_dim_rate (float): (0, 1), default = 0.2, rate of dims will be changed during attraction phase
 
     Examples
     ~~~~~~~~
@@ -32,27 +32,29 @@ class OriginalMGOA(Optimizer):
     >>>     "obj_func": objective_function
     >>> }
     >>>
-    >>> model = MGOA.OriginalMGOA(epoch=5000, pop_size=50)
+    >>> model = MGOA.OriginalMGOA(epoch=1000, pop_size=50, attract_dim_rate=0.2)
     >>> g_best = model.solve(problem_dict)
     >>> print(f"Solution: {g_best.solution}, Fitness: {g_best.target.fitness}")
     >>> print(f"Solution: {model.g_best.solution}, Fitness: {model.g_best.target.fitness}")
 
     References
     ~~~~~~~~~~
-    [1] Symmetry 2025, 17(12), 2118.
+    [1] Liu, S., Xiang, Y., Guo, X., Zhao, F., Zhao, A., & Wu, W. (2025). Market Game Optimization Algorithm:
+    A Metaheuristic Inspired by Symmetric Competitive Behavior of Merchants and Consumers. Symmetry, 17(12), 2118.
     """
 
-    def __init__(self, epoch: int = 5000, pop_size: int = 50, **kwargs: object) -> None:
+    def __init__(self, epoch: int = 5000, pop_size: int = 50, attract_dim_rate=0.2, **kwargs: object) -> None:
         """
         Args:
             epoch (int): maximum number of iterations, default = 5000
             pop_size (int): number of population size, default = 50
+            attract_dim_rate (float): number of dims will be changed in attraction phase under ratio format, default = 0.2
         """
         super().__init__(**kwargs)
         self.epoch = self.validator.check_int("epoch", epoch, [1, 100000])
         self.pop_size = self.validator.check_int("pop_size", pop_size, [5, 10000])
-        self.set_parameters(["epoch", "pop_size"])
-        self.is_parallelizable = False
+        self.attract_dim_rate = self.validator.check_float("attract_dim_rate", attract_dim_rate, (0, 1))
+        self.set_parameters(["epoch", "pop_size", "attract_dim_rate"])
         self.sort_flag = False
 
     def evolve(self, epoch):
@@ -62,35 +64,31 @@ class OriginalMGOA(Optimizer):
         Args:
             epoch (int): The current iteration
         """
-        A = np.pi
         n = 2.0
         alpha = 1.0
         ratio = epoch / self.epoch
-        S = A * np.sin(2 * np.pi * (ratio * n)) * np.exp(-alpha * ratio)
-
+        S = np.pi * np.sin(2 * np.pi * (ratio * n)) * np.exp(-alpha * ratio)
         phase = "attraction" if self.generator.random() < 0.5 else "collaboration"
-        k = 5
+        n_attract_dims = int(np.ceil(self.attract_dim_rate * self.problem.n_dims))
 
+        pop_new = []
         for idx in range(0, self.pop_size):
-            pos = self.pop[idx].solution
-            pos_new = pos.copy()
-
+            pos_new = self.pop[idx].solution.copy()
             if phase == "attraction":
-                n_select = min(self.problem.n_dims, k)
-                selected_dims = self.generator.choice(self.problem.n_dims, n_select, replace=False)
-                pos_new[selected_dims] = pos[selected_dims] - S * (self.g_best.solution[selected_dims] - pos[selected_dims])
+                selected_dims = self.generator.choice(self.problem.n_dims, n_attract_dims, replace=False)
+                pos_new[selected_dims] = pos_new[selected_dims] - S * (self.g_best.solution[selected_dims] - pos_new[selected_dims])
             else:
-                candidates = np.arange(self.pop_size)
-                candidates = candidates[candidates != idx]
-                rA, rB = self.generator.choice(candidates, 2, replace=False)
-                dm1 = self.pop[rA].solution - pos
-                dm2 = self.pop[rB].solution - pos
-                r1 = self.generator.random()
-                r2 = self.generator.random()
-                pos_new = pos + r1 * dm1 + r2 * dm2
+                rA, rB = self.sample_indexes_exclude_one(self.generator, self.pop_size, idx, n_samples=2)
+                dm1 = self.pop[rA].solution - self.pop[idx].solution
+                dm2 = self.pop[rB].solution - self.pop[idx].solution
+                pos_new = self.pop[idx].solution + self.generator.random() * dm1 + self.generator.random() * dm2
 
-            pos_new = np.clip(pos_new, self.problem.lb, self.problem.ub)
             pos_new = self.correct_solution(pos_new)
-            agent = self.generate_agent(pos_new)
-            if self.compare_target(agent.target, self.pop[idx].target, self.problem.minmax):
-                self.pop[idx] = agent
+            agent_new = self.generate_empty_agent(pos_new)
+            pop_new.append(agent_new)
+            if self.mode not in self.AVAILABLE_MODES:
+                agent_new.target = self.get_target(pos_new)
+                self.pop[idx] = self.get_better_agent(self.pop[idx], agent_new, minmax=self.problem.minmax)
+        if self.mode in self.AVAILABLE_MODES:
+            pop_new = self.update_target_for_population(pop_new)
+            self.pop = self.greedy_selection_population(self.pop, pop_new, self.problem.minmax)